Adaptations to changes in GDL.
git-svn-id: svn://localhost@531 8062f311-0dae-4547-b526-b8ab9ac864a5

+++ Extensions/TimedEvent/TimedEventTools.cpp
@@ -33,7 +33,7 @@ namespace GDpriv
 namespace TimedEvents
 {
 
-float GD_EXTENSION_API UpdateAndGetTimeOf(RuntimeScene & scene, std::string timedEventName)
+unsigned int GD_EXTENSION_API UpdateAndGetTimeOf(RuntimeScene & scene, std::string timedEventName)
 {
     TimedEventsManager & manager = TimedEventsManager::managers[&scene];
     manager.timedEvents[timedEventName].UpdateTime(scene.GetElapsedTime());
